Separate UGC components into individual files
