+++ components/UGC/Categories/index.js
@@ -1,0 +1,13 @@
+import CategoryFour from "@/components/Category/CategoryFour";
+
+const Categories = () => (
+  <div className="rbt-categories-area bg-color-extra2 rbt-section-gap">
+    <div className="container">
+      <div className="row g-5">
+        <CategoryFour />
+      </div>
+    </div>
+  </div>
+);
+
+export default Categories;
