+++ components/UGC/WhyChooseUs/index.js
@@ -1,0 +1,22 @@
+import Counter from "@/components/Counters/Counter";
+
+const WhyChooseUs = () => (
+  <div className="rbt-counterup-area rbt-section-gapBottom bg-gradient-1">
+    <div className="container">
+      <div className="row mb--60">
+        <div className="col-lg-12">
+          <div className="section-title text-center">
+            <span className="subtitle bg-primary-opacity">Why Choose Us</span>
+            <h2 className="title">Why Choose Histudy</h2>
+            <p className="description has-medium-font-size mt--20 mb--0">
+              Language Academic?
+            </p>
+          </div>
+        </div>
+      </div>
+      <Counter head={false} />
+    </div>
+  </div>
+);
+
+export default WhyChooseUs;
